Project Finished
All 4 GUIs complete.
GUI and server are 2 separate programs.
Code assumes SQL database server is setup.
Database Models folder contains the schema used as well as some default
data initiation for the database. The default clientID is 1 for this
data.
For those wishing to test code download mysql server and mysql workbench
and apply the schema and data to a server.
Code for GUI and server is Java, projects created using netbeans.
Some example code was adapted and was referenced in the respective
files.